# contracts/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "AiNexus.Api | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5181/"
    }
  ],
  "paths": {
...
    "/api/v1/models": {
      "get": {
        "tags": ["Inference"],
        "operationId": "ListModels",
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelsDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/context": {
      "post": {
        "tags": ["Inference"],
        "operationId": "PreviewContext",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContextPreviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContextUsageDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/runs": {
      "post": {
        "tags": ["Inference"],
        "operationId": "CreateRun",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "202": {
            "description": "Accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/runs/{id}": {
      "get": {
        "tags": ["Inference"],
        "operationId": "GetRun",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/runs/{id}/cancel": {
      "post": {
        "tags": ["Inference"],
        "operationId": "CancelRun",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/runs/{id}/events": {
      "get": {
        "tags": ["Inference"],
        "operationId": "RunEvents",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "after",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "text/event-stream": {
                "schema": {
                  "$ref": "#/components/schemas/RunEventDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ContextPreviewRequest": {
        "required": ["conversationId", "parentMessageId", "prompt", "modelId"],
        "type": "object",
        "properties": {
          "conversationId": {
            "type": ["null", "string"],
            "format": "uuid"
          },
          "parentMessageId": {
            "type": ["null", "string"],
            "format": "uuid"
          },
          "prompt": {
            "type": ["null", "string"]
          },
          "modelId": {
            "type": ["null", "string"]
          },
          "attachmentIds": {
            "type": ["null", "array"],
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        }
      },
      "ContextUsageDto": {
        "required": [
          "estimatedInputTokens",
          "contextTokens",
          "reservedOutputTokens",
          "droppedMessages",
          "budgetExceeded"
        ],
        "type": "object",
        "properties": {
          "estimatedInputTokens": {
            "type": "integer",
            "format": "int32"
          },
          "contextTokens": {
            "type": "integer",
            "format": "int32"
          },
          "reservedOutputTokens": {
            "type": "integer",
            "format": "int32"
          },
          "droppedMessages": {
            "type": "integer",
            "format": "int32"
          },
          "budgetExceeded": {
            "type": "boolean"
          },
          "isEstimate": {
            "type": "boolean",
            "default": true
          },
          "reservedKnowledgeTokens": {
            "type": "integer",
            "format": "int32",
            "default": 0
          }
        }
      },
...
      "CreateRunRequest": {
        "required": [
          "conversationId",
          "modelId",
          "prompt",
          "parentMessageId",
          "regenerateUserMessageId"
        ],
        "type": "object",
        "properties": {
          "conversationId": {
            "type": "string",
            "format": "uuid"
          },
          "modelId": {
            "type": ["null", "string"]
          },
          "prompt": {
            "type": ["null", "string"]
          },
          "parentMessageId": {
            "type": ["null", "string"],
            "format": "uuid"
          },
          "regenerateUserMessageId": {
            "type": ["null", "string"],
            "format": "uuid"
          },
          "reasoningEffort": {
            "type": ["null", "string"]
          },
          "attachmentIds": {
            "type": ["null", "array"],
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        }
      },
...
      "ModelDto": {
        "required": [
          "id",
          "displayName",
          "contextTokens",
          "maxOutputTokens",
          "supportsStreaming",
          "supportsUsage",
          "reasoningEfforts",
          "defaultReasoningEffort"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "contextTokens": {
            "type": "integer",
            "format": "int32"
          },
          "maxOutputTokens": {
            "type": "integer",
            "format": "int32"
          },
          "supportsStreaming": {
            "type": "boolean"
          },
          "supportsUsage": {
            "type": "boolean"
          },
          "reasoningEfforts": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "defaultReasoningEffort": {
            "type": "string"
          },
          "supportsImages": {
            "type": "boolean",
            "default": false
          }
        }
      },
      "ModelPolicyDto": {
        "required": ["allowModelSelection", "showModelNames", "defaultModelId"],
        "type": "object",
        "properties": {
          "allowModelSelection": {
            "type": "boolean"
          },
          "showModelNames": {
            "type": "boolean"
          },
          "defaultModelId": {
            "type": ["null", "string"]
          },
          "maxInputCharacters": {
            "type": "integer",
            "format": "int32",
            "default": 12000
          }
        }
      },
      "ModelsDto": {
        "required": ["models", "providerAvailable", "notice", "policy"],
        "type": "object",
        "properties": {
          "models": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ModelDto"
            }
          },
          "providerAvailable": {
            "type": "boolean"
          },
          "notice": {
            "type": ["null", "string"]
          },
          "policy": {
            "$ref": "#/components/schemas/ModelPolicyDto"
          }
        }
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": ["null", "string"]
          },
          "title": {
            "type": ["null", "string"]
          },
          "status": {
            "type": ["null", "integer"],
            "format": "int32"
          },
          "detail": {
            "type": ["null", "string"]
          },
          "instance": {
            "type": ["null", "string"]
          }
        }
      },
...
      "RunDto": {
        "required": [
          "id",
          "conversationId",
          "userMessageId",
          "assistantMessageId",
          "modelId",
          "status",
          "content",
          "lastSequence",
          "errorCode",
          "createdAt",
          "startedAt",
          "finishedAt",
          "inputTokens",
          "outputTokens"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "conversationId": {
            "type": "string",
            "format": "uuid"
          },
          "userMessageId": {
            "type": "string",
            "format": "uuid"
          },
          "assistantMessageId": {
            "type": "string",
            "format": "uuid"
          },
          "modelId": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "lastSequence": {
            "type": "integer",
            "format": "int64"
          },
          "errorCode": {
            "type": ["null", "string"]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "startedAt": {
            "type": ["null", "string"],
            "format": "date-time"
          },
          "finishedAt": {
            "type": ["null", "string"],
            "format": "date-time"
          },
          "inputTokens": {
            "type": ["null", "integer"],
            "format": "int64"
          },
          "outputTokens": {
            "type": ["null", "integer"],
            "format": "int64"
          }
        }
      },
      "RunEventDto": {
        "required": [
          "version",
          "sequence",
          "runId",
          "type",
          "status",
          "delta",
          "errorCode"
        ],
        "type": "object",
        "properties": {
          "version": {
            "type": "integer",
            "format": "int32"
          },
          "sequence": {
            "type": "integer",
            "format": "int64"
          },
          "runId": {
            "type": "string",
            "format": "uuid"
          },
          "type": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "delta": {
            "type": ["null", "string"]
          },
          "errorCode": {
            "type": ["null", "string"]
          }
        }
      },
...
    }
  },
...
}